seguranca_software/r1/pyproject.toml
[project]
name = "r1"
version = "0.1.0"
description = ""
authors = [
    {name = "V.Guilherme",email = "v.guilherme.barreto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
